Seperate wan_backbone to WAM_BACKBONES
