+++ fluxvla/models/backbones/__init__.py
@@ -15,3 +15,4 @@
 from .llms import *  # noqa: F401, F403
 from .visions import *  # noqa: F401, F403
 from .vlms import *  # noqa: F401, F403
+from .wams import *  # noqa: F401, F403
