+++ fluxvla/models/backbones/wams/__init__.py
@@ -12,20 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .registry import Registry
-
-TOKENIZERS = Registry('tokenizers')
-TRANSFORMS = Registry('transforms')
-DATASETS = Registry('datasets')
-LLM_BACKBONES = Registry('llm_backbones')
-VISION_BACKBONES = Registry('vision_backbones')
-PROJECTORS = Registry('projectors')
-HEADS = Registry('heads')
-VLAS = Registry('vlas')
-RUNNERS = Registry('runners')
-COLLATORS = Registry('collators')
-METRICS = Registry('metrics')
-PROCESSORS = Registry('processors')
-VLM_BACKBONES = Registry('vlm_backbones')
-WAM_BACKBONES = Registry('wam_backbones')
-OPERATORS = Registry('operators')
+try:
+    from .wan_backbone import WanBackbone  # noqa: F401
+except ImportError:
+    pass
